added observable-data-summary #599
